Split codebase into separate packages for core and each language type
