+++ src/main/scala/bridges/core/Declaration.scala
@@ -1,3 +1,3 @@
-package bridges
+package bridges.core
 
 final case class Declaration(id: String, tpe: Type)
