+++ src/main/scala/bridges/core/JsonDecoder.scala
@@ -1,0 +1,9 @@
+package bridges.core
+
+import unindent._
+
+trait JsonDecoder[A] {
+  def decoder(decl: Declaration): String
+  def decoder(decls: List[Declaration]): String =
+    decls.map(decoder).mkString("\n\n")
+}
