+++ src/main/scala/bridges/core/JsonEncoder.scala
@@ -1,0 +1,9 @@
+package bridges.core
+
+import unindent._
+
+trait JsonEncoder[A] {
+  def encoder(decl: Declaration): String
+  def encoder(decls: List[Declaration]): String =
+    decls.map(encoder).mkString("\n\n")
+}
