+++ src/main/scala/bridges/elm/Elm.scala
@@ -1,0 +1,10 @@
+package bridges.elm
+
+import bridges.core._
+
+sealed trait Elm
+
+object Elm extends ElmJsonDecoder
+  with ElmJsonEncoder
+  with ElmRenderer
+  with ElmFileBuilder
