+++ src/main/scala/bridges/flow/Flow.scala
@@ -1,0 +1,5 @@
+package bridges.flow
+
+sealed trait Flow
+
+object Flow extends FlowRenderer
