+++ src/main/scala/bridges/typescript/Typescript.scala
@@ -1,0 +1,5 @@
+package bridges.typescript
+
+sealed trait Typescript
+
+object Typescript extends TypescriptRenderer
